requirements.txt
beautifulsoup4==4.13.3
openai==1.70.0
openpyxl==3.1.5
PyPDF2==3.0.1
python-dotenv==1.1.0
python_docx==1.1.2
python_pptx==1.0.2
streamlit==1.44.1
